Fix comparison type

+++ libtac/lib/attrib.c
@@ -20,6 +20,8 @@
  * See `CHANGES' file for revision history.
  */
 
+#include <stdio.h>
+
 #include "libtac.h"
 #include "xalloc.h"
 
@@ -109,7 +111,7 @@ static int _tac_add_attrib_pair(struct tac_attrib **attr, char *name,
 
     a->attr = (char *)xcalloc(1, total_len + 1);
     /* write the attribute=value into the buffer */
-    if (snprintf(a->attr, total_len, "%s%c%s", name, sep, value) < (size_t) total_len)
+    if (snprintf(a->attr, total_len, "%s%c%s", name, sep, value) < (int) total_len)
     {
         TACSYSLOG(LOG_ERR, 
                       "%s: short snprintf write (wanted %lu bytes)",
